Fixed FAD Bugs
Fixed a bug that cut off the last probability block. Fixed the centering process: now the mean is subtracted from all values so that the distribution is centered at 0 standard deviations from the mean. The FAD generation function should now be capable of giving good estimations of arbitrary distributions, but more testing is necessary.